refactor: Audio modules to core.
