+++ src/core/audio/musicplayer.js
